ML-VVSU-2025_5.1/requirements.txt
beautifulsoup4>=4.12.0
lxml>=4.9.0
requests>=2.31.0
python-dateutil>=2.8.0
